refactor(modals): rename to ModalHeader and allow optional durations
